requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
scipy>=1.7.0
matplotlib>=3.4.0
PySide6>=6.5.0
torch>=2.3.0
torchvision>=0.18.0
tensorboard>=2.11.0
